# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.Response:
    properties:
      message:
        type: string
    type: object
  inputs.Airport:
    properties:
      address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  inputs.CreateFlight:
    properties:
      arrival:
        $ref: '#/definitions/inputs.Airport'
      departure:
        $ref: '#/definitions/inputs.Airport'
    required:
    - arrival
    - departure
    type: object
...
  inputs.CreateVehicle:
    properties:
      cruise_altitude:
        type: number
      cruise_speed:
        type: number
      is_verified:
        type: boolean
      matriculation:
        type: string
      model_name:
        type: string
      seat:
        type: integer
      type:
        type: string
    required:
    - cruise_altitude
    - cruise_speed
    - matriculation
    - model_name
    - seat
    - type
    type: object
  inputs.InputCreateProposal:
    properties:
      availableSeats:
        type: integer
      createFlight:
        $ref: '#/definitions/inputs.CreateFlight'
      departureTime:
        type: string
      description:
        type: string
      price:
        type: number
      vehicleId:
        type: integer
    required:
    - availableSeats
    - createFlight
    - departureTime
    - description
    - price
    - vehicleId
    type: object
...
  inputs.UpdateVehicle:
    properties:
      cruise_altitude:
        type: number
      cruise_speed:
        type: number
      is_selected:
        type: boolean
      is_verified:
        type: boolean
      matriculation:
        type: string
      model_name:
        type: string
      seat:
        type: integer
      type:
        type: string
      user_id:
        type: integer
    type: object
  responses.ListUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      last_name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.ResponseAirport:
    properties:
      address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  responses.ResponseFlight:
    properties:
      arrival:
        $ref: '#/definitions/responses.ResponseAirport'
      created_at:
        type: string
      departure:
        $ref: '#/definitions/responses.ResponseAirport'
      id:
        type: integer
      pilot:
        $ref: '#/definitions/responses.ListUser'
      pilot_id:
        type: integer
      price:
        type: number
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      vehicle:
        $ref: '#/definitions/responses.Vehicle'
      vehicle_id:
        type: integer
    required:
    - arrival
    - departure
    - id
    - status
    type: object
  responses.ResponseProposal:
    properties:
      available_seats:
        type: integer
      departure_time:
        type: string
      description:
        type: string
      flight:
        $ref: '#/definitions/responses.ResponseFlight'
      id:
        type: integer
    type: object
...
  responses.Vehicle:
    properties:
      created_at:
        type: string
      cruise_altitude:
        type: number
      cruise_speed:
        type: number
      id:
        type: integer
      is_selected:
        type: boolean
      is_verified:
        type: boolean
      matriculation:
        type: string
      model_name:
        type: string
      seat:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: this is the go project.
  title: Backend AirFleet
  version: "1.0"
paths:
...
  /proposals:
    get:
      consumes:
      - application/json
      description: Get all proposals
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.ResponseProposal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Get all proposals
      tags:
      - proposals
    post:
      consumes:
      - application/json
      description: Create a proposal
      parameters:
      - description: Proposal
        in: body
        name: proposal
        required: true
        schema:
          $ref: '#/definitions/inputs.InputCreateProposal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseProposal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Create a proposal
      tags:
      - proposals
  /proposals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a proposal
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Delete a proposal
      tags:
      - proposals
    get:
      consumes:
      - application/json
      description: Get a proposal
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseProposal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Get a proposal
      tags:
      - proposals
...
  /vehicles:
    get:
      consumes:
      - application/json
      description: get all vehicles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Vehicle'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Get all vehicles
      tags:
      - vehicle
    post:
      consumes:
      - application/json
      description: create a new vehicle for a pilot
      parameters:
      - description: Message body
        in: body
        name: vehicleInput
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateVehicle'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Vehicle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Create vehicle
      tags:
      - vehicle
  /vehicles/{id}:
    delete:
      consumes:
      - application/json
      description: delete vehicle by id
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: delete vehicle by id
      tags:
      - vehicle
    get:
      consumes:
      - application/json
      description: get vehicle by id
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Vehicle'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: get vehicle by id
      tags:
      - vehicle
    patch:
      consumes:
      - application/json
      description: update vehicle by id
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message body
        in: body
        name: vehicleInput
        required: true
        schema:
          $ref: '#/definitions/inputs.UpdateVehicle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Vehicle'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: update vehicle by id
      tags:
      - vehicle
  /vehicles/me:
    get:
      consumes:
      - application/json
      description: get all vehicles for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Vehicle'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
      security:
      - BearerAuth: []
      summary: Get all vehicles for current user
      tags:
      - vehicle
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"